Next snapshot

+++ hornet-mainnet-k8s/hornet-k8s.sh
@@ -8,7 +8,7 @@
 set -e
 
 help () {
-  echo "usage: hornet-k8s.sh [deploy|scale|update|undeploy]"
+  echo "usage: hornet-k8s.sh [deploy|scale|undeploy]"
   echo "Parameter: NAMESPACE=<Kubernetes Namespace>"
   echo "Parameter: INSTANCES=<Number of Instances>"
   echo "Parameter: PEER=<multiPeerAddress>"
@@ -20,16 +20,18 @@ help () {
 command="$1"
 peer="$PEER"
 namespace="$NAMESPACE"
-declare -i instances="$INSTANCES"
+declare -i instances=1
 declare -i replicas=0
 ingress_class="$INGRESS_CLASS"
+
+configmap_name="hornet-config"
 
 if [ -z "$namespace" ]; then
     namespace="tangle"
 fi
 
-if [ -z "$instances" ]; then
-    instances=1
+if ! [ -z "$INSTANCES" ]; then
+    instances=$INSTANCES
 fi
 
 if [ -z "$ingress_class" ]; then
@@ -84,20 +86,27 @@ createSecret () {
     
     kubectl  -n $namespace create secret generic hornet-secret --from-literal='DASHBOARD_AUTH_PASSWORDHASH='"$dashboard_hash" \
     --from-literal='DASHBOARD_AUTH_PASSWORDSALT='"$dashboard_salt" --dry-run=client -o yaml | kubectl apply -f -
+}
 
+createPrivateKeys () {
     mkdir -p config/keys
 
     # for each of the instances a new secret with private key is created
-    for  (( i=$replicas; i<$instances; i++ ))
+    for  (( i=0; i<instances; i++ ))
     do
-        # If the proper version of OpenSSL is not installed this has highly chances fo failure
-        set +e
-        openssl genpkey -algorithm ed25519 -out config/keys/identity.key
-        if ! [ $? -eq 0 ]; then
-            echo "Cannot generate a private key for your node. Please check your OpenSSL version"
-            exit -100
+        # If they key already exists we just reuse it (normally during an scale operation)
+        if ! [ -f config/keys/identity-$i.key ]; then
+            # If the proper version of OpenSSL is not installed this has highly chances fo failure
+            set +e
+            openssl genpkey -algorithm ed25519 -out config/keys/identity.key
+            if ! [ $? -eq 0 ]; then
+                echo "Cannot generate a private key for your node. Please check your OpenSSL version"
+                exit -100
+            fi
+            set -e
+        else
+            mv config/keys/identity-$i.key config/keys/identity.key 
         fi
-        set -e
 
         # We need to store this to later perform the Node peering
         p2p_identity=$(docker run -it -v "$PWD/config/keys:/p2pstore/" gohornet/hornet:1.1.3 tool p2pidentity-extract /p2pstore |\
@@ -118,7 +127,7 @@ createStatefulSet () {
 }
 
 createNodePortServices () {
-    for  (( i=$replicas; i<$instances; i++ ))
+    for  (( i=replicas; i<instances; i++ ))
     do
         cat hornet-service.yaml | kubectl patch --dry-run=client -p \
         $'metadata:\n  namespace: '"$namespace" -o yaml -f - \
@@ -131,8 +140,12 @@ createNodePortServices () {
 
 # Peers the nodes between them
 peerNodes () {
-    local start=$((replicas+1))
-    for  (( i=$start; i<$instances; i++ ))
+    declare -i start=1
+    if [ $replicas -gt 0 ]; then
+        start=replicas
+    fi
+
+    for  (( i=start; i<instances; i++ ))
     do
        peerHost=hornet-$((i-1))
        peerAddr=${p2p_identities[$((i-1))]}
@@ -142,8 +155,10 @@ peerNodes () {
        jq '.' | jq '.peers |= . + [{ "alias": "'$alias'", "multiAddress": "'$multiAddress'"}]' > config/peering-$i.json
     done
 
-    # Peering config of the first node
-    mv config/peering.json config/peering-0.json
+    if [ -f config/peering.json ]; then
+        # Peering config of the first node
+        mv config/peering.json config/peering-0.json
+    fi
 
     echo "Nodes peered"
 }
@@ -172,6 +187,7 @@ deployHornet () {
     kubectl create namespace $namespace --dry-run=client -o yaml | kubectl apply -f -
 
     createSecret
+    createPrivateKeys
 
     cooSetup
 
@@ -180,8 +196,8 @@ deployHornet () {
     # Now we peer the nodes among themselves
     peerNodes
 
-    # Config Map is created or overewritten
-    kubectl -n $namespace create configmap hornet-config --from-file=config --dry-run=client -o yaml | kubectl apply -f -
+    # Config Map is created or overwritten
+    kubectl -n $namespace create configmap $configmap_name --from-file=config --dry-run=client -o yaml | kubectl apply -f -
 
     # Service, Ingress associated and Statefulset associated
     kubectl apply -n $namespace -f hornet-rest-service.yaml
@@ -205,15 +221,27 @@ undeployHornet () {
 }
 
 scaleHornet () {
+    # In order to scale we need to check whether the set already exists
+    set +e
+    kubectl get statefulsets/hornet-set -n $namespace > /dev/null 2>&1
+    if ! [ $? -eq 0 ]; then
+        echo "Error: Statefulset 'hornet-set' does not exist on namespace '$namespace'. Deploy first"
+        exit 400
+    fi
+    set -e
+
     # First we need to know whether we are upscaling or downscaling
     # If upscaling we need to generate additional secrets and additional peering configuration
     replicas=$(kubectl get statefulsets/hornet-set -n $namespace -o jsonpath='{.spec.replicas}' | tr -d '\r\n')
-    if [ $instances -gt $replicas]; then
+    if [ $instances -gt $replicas ]; then
         # We create additional secrets if needed
-        createSecret
+        createPrivateKeys
          # Now we peer the nodes among themselves
         peerNodes
     fi
+
+    # Config Map has to be recreated so that the new peering info is provided
+    kubectl -n $namespace create configmap $configmap_name --from-file=config --dry-run=client -o yaml | kubectl apply -f -
 
     kubectl scale -n $namespace statefulsets hornet-set --replicas=$instances
     # Finally the NodePort services are created
